+++ tests/sdk/query/test_having.py
@@ -36,7 +36,7 @@ class TestBaseTypes:
     def test_queries(self, test_databases):
         query = "SELECT age, sex, COUNT(*) AS n, SUM(income) AS income FROM PUMS.PUMS GROUP BY age, sex HAVING income > 100000"
         privacy = Privacy(10.0, 10E-3)
-        readers = test_databases.create_private_readers(metadata=pums_schema_path, privacy=privacy, database='PUMS')
+        readers = test_databases.create_private_readers(privacy=privacy, database='PUMS')
         for reader in readers:
             res = [len(self.reader.execute(query)) for i in range(5)]
             assert np.mean(res) < 115 and np.mean(res) > 10 # actual is 14, but noise is huge
